+++ src/lib/components/layout/NavBar.svelte
@@ -159,4 +159,3 @@
 		border-color: transparent var(--color-accent-900) transparent transparent;
 	}
 </style>
-