+++ src/lib/components/prestige/AtomRealm.svelte
@@ -6,27 +6,31 @@
 	import Canvas from '@components/game/Canvas.svelte';
 	import Counter from '@components/game/Counter.svelte';
 	import Upgrades from '@components/game/Upgrades.svelte';
-	import { mobile } from '$stores/window.svelte';
+	import { RealmTypes } from '$data/realms';
 	import { gameManager } from '$helpers/GameManager.svelte';
+	import { realmManager } from '$helpers/RealmManager.svelte';
+	import { mobile } from '$stores/window.svelte';
 
 	let activeTab: 'achievements' | 'buildings' | 'upgrades' = $state('upgrades');
 </script>
 
-<div class="relative pt-12 transition-all duration-1000 ease-in-out lg:pt-4 {mobile.current ? 'min-h-screen pb-8' : ''}">
-	<div class="-z-10 absolute inset-0 overflow-hidden pointer-events-none">
-		{#if gameManager.totalProtonisesAllTime > 0}
-			<div class="absolute bg-yellow-400/15 blur-[160px] h-64 right-[20%] rounded-full top-[10%] w-64"></div>
-		{/if}
-		{#if gameManager.totalElectronizesAllTime > 0}
-			<div class="absolute bg-green-500/15 blur-[180px] bottom-[20%] h-80 left-[10%] rounded-full w-80"></div>
-		{/if}
-	</div>
-
+<div class="relative pt-12 transition-all duration-1000 ease-in-out lg:pt-8 {mobile.current ? 'min-h-screen pb-8' : ''}">
 	<Canvas />
+
+	{#if realmManager.selectedRealmId === RealmTypes.ATOMS}
+		<div class="fixed inset-0 -z-50 pointer-events-none overflow-hidden">
+			{#if gameManager.totalProtonisesAllTime > 0}
+				<div class="absolute bg-yellow-400/15 blur-[160px] h-64 right-[20%] rounded-full top-[10%] w-64"></div>
+			{/if}
+			{#if gameManager.totalElectronizesAllTime > 0}
+				<div class="absolute bg-green-500/15 blur-[180px] bottom-[10%] h-80 left-[10%] rounded-full w-80"></div>
+			{/if}
+		</div>
+	{/if}
 	<Bonus />
 
 	<div class="game-container gap-8 grid lg:max-w-4xl mx-auto p-8 text-sm xl:max-w-360">
-		<div class="left-panel flex flex-col gap-4 z-10">
+		<div class="left-panel flex flex-col gap-1.5 z-10">
 			<div class="grid grid-flow-col gap-2 auto-cols-fr">
 				<button
 					class="backdrop-blur-xs rounded-lg p-2 w-full whitespace-nowrap border-none text-inherit cursor-pointer transition-all duration-200 {activeTab === 'upgrades'
@@ -51,7 +55,7 @@
 					Achievements
 				</button>
 			</div>
-			<div class="flex-1 overflow-y-auto {mobile.current ? 'max-h-[60vh]' : ''}">
+			<div class="mt-1">
 				{#if activeTab === 'upgrades'}
 					<Upgrades />
 				{:else if activeTab === 'achievements'}
@@ -66,7 +70,7 @@
 			<Atom />
 		</div>
 		{#if !mobile.current}
-			<div class="right-panel">
+			<div class="right-panel pt-12">
 				<Buildings />
 			</div>
 		{/if}